+++ src/main/resources/assets/cursedmod/lang/en_us.json
@@ -1,0 +1,10 @@
+{
+  "item.cursedmod.vjb": "VJB",
+
+  "itemGroup.cursedmod.vjbgroup": "cursed VJB",
+
+  "block.cursedmod.vjblock": "VJBLOCK",
+  "block.cursedmod.vjbore": "VJBORE",
+
+  "enchantment.cursedmod.son": "Musique"
+}
